requirements.txt
python-dotenv
streamlit
PyPDF2
docx
boto3
opensearch-py 
elasticsearch